Add Powershell Capability
This provides a PowerShell function for performing scans.  Make sure to use the PowerShell folder for the yaml files.  Some minor changes were made to a small selection of regexes as PowerShell struggled with some of them.   Was mostly in the trufflehog section if I remember correctly. I've tested the function today across 90 files with fairly high line counts and it performs well.